requirements.txt
langchain-google-genai
python-dotenv
langchain_experimental
psycopg2-binary
fastapi 
uvicorn 
langchain 
